requirements.txt
# PDF 解析
PyMuPDF>=1.22.0
pdfplumber>=0.9.0
pikepdf>=4.4.0

# OCR 引擎
rapidocr-onnxruntime>=1.3.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Elasticsearch 客户端
elasticsearch>=8.10.0

# NLP 与中文处理
jieba>=0.42.1

# 图像处理
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0,<2.0.0

# 工具库
tqdm>=4.60.0
pydantic>=2.0.0
python-dotenv>=1.0.0
